find middle ground with separating components
